GetMoviesListController

+++ models/movie.go
@@ -14,7 +14,7 @@ type MovieData struct {
 	Name string `json:"name"`
 	ReleaseDate string `json:"release_date"`
 	OpeningCrawl string `json:"opening_crawl"`
-	CommentCount int `json:"comment_count"`
+	CommentCount int64 `json:"comment_count"`
 }
 
 type SomeMovie struct {
